Move `build_docs` and `upload_products` commands into `_setup` module
