+++ setup.py
@@ -3,10 +3,12 @@ import os
 from distutils.core import Command
 import versioneer
 
-from _setup.efficiency_test import efficiency_test
-from _setup.copy_script import copy_script
+from _setup.build_docs import build_docs
 from _setup.changelog import changelog
+from _setup.copy_script import copy_script
+from _setup.efficiency_test import efficiency_test
 from _setup.testClass import _TestClass
+from _setup.upload_products import upload_products
 
 # bootstrap setuptools for users that don't already have it
 import ez_setup
@@ -31,133 +33,7 @@ except ImportError, e:
 
 
 
-class build_docs(Command):
-
-    description = "build the FiPy documentation"
-
-    # List of option tuples: long name, short name (None if no short
-    # name), and help string.
-    user_options = [('pdf', None, "compile the PDF variant of the documentation"),
-                    ('html', None, "compile the HTML variant of the documentation"),
-                    ('cathartic', None, "rewrite all the files (default is to only rewrite changed files)"),
-                   ]
-
-    def initialize_options (self):
-        self.pdf = 0
-        self.html = 0
-        self.cathartic = 0
-
-    def finalize_options (self):
-        pass
-
-    def run (self):
-        import sphinx.cmd.build
-        import sphinx.ext.apidoc
-        
-        sphinx_args = ['-P', '-n', '-c', 'documentation/', '.']
-        apidoc_args = []
-        
-        if self.cathartic:
-            sphinx_args = ['-a', '-E'] + sphinx_args
-            apidoc_args = ['--force'] + apidoc_args
             
-        sphinx.ext.apidoc.main(['--output-dir=fipy/generated', '--suffix=rst']
-                    + apidoc_args + ['fipy'])
-        sphinx.ext.apidoc.main(['--output-dir=documentation/tutorial/package/generated', '--suffix=rst']
-                    + apidoc_args + ['documentation/tutorial/package'])
-
-        if self.html:
-            sphinx.cmd.build.main(['-b', 'redirecting_html'] + sphinx_args + ['documentation/_build/html/'])
-
-        if self.pdf:
-            try:
-                sphinx.cmd.build.main(['-b', 'latex'] + sphinx_args + ['documentation/_build/latex/'])
-            except SystemExit:
-                pass
-            
-            outdir = os.path.join('documentation', '_build', 'latex')
-            
-            from docutils.core import publish_file
-
-            for xtra in ("LICENSE", "DISCLAIMER"):
-                publish_file(source_path="%s.rst" % xtra,
-                             destination_path=os.path.join(outdir, "%s.tex" % xtra),
-                             reader_name='standalone',
-                             parser_name='restructuredtext',
-                             writer_name='latex',
-                             settings_overrides= {
-                                 'template': 'documentation/_templates/empty.tex'
-                             })
-
-            savedir = os.getcwd()
-            
-            os.chdir(outdir)
-                
-            os.system("pdflatex fipy")
-            os.system("pdflatex fipy")
-            os.system("pdflatex fipy")
-            os.system("makeindex -s python.ist fipy")
-            os.system("makeindex -s python.ist modfipy")
-            os.system("pdflatex fipy")
-            os.system("pdflatex fipy")
-                
-            os.chdir(savedir)
-            
-class upload_products(Command):
-    description = "upload FiPy compressed archives to website(s)"
-    
-    user_options = [('pdf', None, "upload the PDF variant of the documentation"),
-                    ('html', None, "upload the HTML variant of the documentation"),
-                    ('tarball', None, "upload the .tar.gz source distribution"),
-                    ('winzip', None, "upload the .win32.zip distribution"),
-                   ]
-
-    def initialize_options (self):
-        self.pdf = 0
-        self.html = 0
-        self.tarball = 0
-        self.winzip = 0
-
-    def finalize_options (self):
-        pass
-
-    def run(self):
-        if self.pdf:
-            print "setting permissions of manual..."
-            os.system('chmod -R g+w documentation/_build/latex/fipy.pdf')
-            
-            print "linking manual to `dist/`..."
-            os.system('mkdir dist/')
-            os.system('ln -f documentation/_build/latex/fipy.pdf dist/fipy-%s.pdf'%self.distribution.metadata.get_version())
-            
-        if self.html:
-            print "setting group and ownership of web pages..."
-            os.system('chmod -R g+w documentation/_build/html/')
-            
-            print "uploading web pages..."
-            # The -t flag (implicit in -a) is suddenly causing problems
-            # os.system('rsync -aLC -e ssh %s %s'%('documentation/www/', os.environ['FIPY_WWWHOST']))
-            os.system('rsync -rlpgoDLC -e ssh %s %s' % ('documentation/_build/html/', os.environ['FIPY_WWWHOST']))
-
-        if self.tarball:
-            file = 'dist/FiPy-%s.tar.gz' % self.distribution.metadata.get_version()
-            print "setting permissions for %s ..." % file
-            os.system('chmod -R g+w %s' % file)
-
-            print "uploading tarball..."
-            os.system('rsync -pgoDLC -e ssh %s %s/download/' % (file, os.environ['FIPY_WWWHOST']))
-
-        if self.winzip:
-            file = 'dist/FiPy-%s.win32.zip' % self.distribution.metadata.get_version()
-            print "setting permissions for %s ..." % file
-            os.system('chmod -R g+w %s' % file)
-            
-            print "uploading winzip..."
-            os.system('rsync -pgoDLC -e ssh %s %s/download/' % (file, os.environ['FIPY_WWWHOST']))
-
-        if self.pdf or self.tarball or self.winzip:
-            print "activating web pages..."
-            os.system(os.environ['FIPY_WWWACTIVATE'])
 
 try:            
     f = open('README.rst', 'r')
